[FEATURE] Button Actions

+++ src/main/java/com/seailz/discordjv/utils/discordapi/DiscordRequest.java
@@ -175,9 +175,9 @@ public class DiscordRequest {
                 Logger logger = Logger.getLogger("DiscordJv");
                 logger.warning("[RATE LIMIT] Rate limit exceeded, waiting for " + response.headers().map().get("Retry-After").get(0) + " seconds");
                 djv.getRateLimits().put(baseUrl, new RateLimit(
-                        Integer.parseInt(response.headers().map().get("X-RateLimit-Limit").get(0)),
-                        Integer.parseInt(response.headers().map().get("X-RateLimit-Remaining").get(0)),
-                        Integer.parseInt(response.headers().map().get("X-RateLimit-Reset-After").get(0))
+                        Double.parseDouble(response.headers().map().get("X-RateLimit-Limit").get(0)),
+                        Double.parseDouble(response.headers().map().get("X-RateLimit-Remaining").get(0)),
+                        Double.parseDouble(response.headers().map().get("X-RateLimit-Reset-After").get(0))
                 ));
                 new Thread(() -> {
                     try {
